+++ Localize/lang/strings.json
@@ -167,6 +167,7 @@
   "5rkegy": "Delete Parameter",
   "5ytHcK": "Loading...",
   "6+RwRp": "Hybrid Connector",
+  "61iT9C": "Insert data from previous step (You can also add by typing Alt + / in the editor)",
   "62Ypnr": "Error loading operation data",
   "63/zYN": "Required. The object to check if it is greater than value being compared to.",
   "65FlWi": "(UTC+08:45) Eucla",
@@ -886,6 +887,7 @@
   "_5rkegy.comment": "Delete Button Tooltip Text",
   "_5ytHcK.comment": "Loading text for spinner",
   "_6+RwRp.comment": "Resource group title",
+  "_61iT9C.comment": "Label for button to open dynamic content token picker",
   "_62Ypnr.comment": "label for panel error",
   "_63/zYN.comment": "Required object parameter to compare to in less function",
   "_65FlWi.comment": "Timezone value ",
