+++ libs/designer-ui/src/lib/editor/base/plugins/tokenpickerbutton/assets/fxIcon.tsx
