+++ libs/services/intl/src/compiled-lang/strings.en-XA.json
@@ -2429,6 +2429,20 @@
       "value": "]"
     }
   ],
+  "61iT9C": [
+    {
+      "type": 0,
+      "value": "["
+    },
+    {
+      "type": 0,
+      "value": "Īƞşḗḗřŧ ḓȧȧŧȧȧ ƒřǿǿḿ ƥřḗḗṽīǿǿŭŭş şŧḗḗƥ (Ẏǿǿŭŭ ƈȧȧƞ ȧȧŀşǿǿ ȧȧḓḓ ƀẏ ŧẏƥīƞɠ Ȧŀŧ + / īƞ ŧħḗḗ ḗḗḓīŧǿǿř)"
+    },
+    {
+      "type": 0,
+      "value": "]"
+    }
+  ],
   "62Ypnr": [
     {
       "type": 0,
