+++ libs/services/intl/src/compiled-lang/strings.json
@@ -1085,6 +1085,12 @@
       "value": "Hybrid Connector"
     }
   ],
+  "61iT9C": [
+    {
+      "type": 0,
+      "value": "Insert data from previous step (You can also add by typing Alt + / in the editor)"
+    }
+  ],
   "62Ypnr": [
     {
       "type": 0,
